feat: added user.service and renamed lessons service
